Backend Code Refactor Completed ✅
