+++ routes/comment.js
@@ -10,7 +10,7 @@ import {
 export const commentRouter = express.Router();
 
 commentRouter
-  .get("/add-comment", verifyJWT, addComment)
+  .post("/add-comment", verifyJWT, addComment)
   .post("/get-blog-comments", getBlogComments)
   .post("/get-replies", getReplies)
   .post("/delete-comment", verifyJWT, deleteComment);
