+++ server.js
@@ -6,6 +6,7 @@ import admin from "firebase-admin";
 import { authRouter } from "./routes/auth.js";
 import { blogRouter } from "./routes/blog.js";
 import { userRouter } from "./routes/user.js";
+import { commentRouter } from "./routes/comment.js";
 import { notificationRouter } from "./routes/notification.js";
 import serviceAccountKey from "./mordern-blogging-platfrom-firebase-adminsdk-et5e8-0590012c08.json" assert { type: "json" };
 
@@ -26,7 +27,7 @@ mongoose.connect(process.env.DB_LOCATION, {
 app.use("/auth", authRouter);
 app.use("/user", userRouter);
 app.use("/blog", blogRouter);
-app.use("/comment", blogRouter);
+app.use("/comment", commentRouter);
 app.use("/notifications", notificationRouter);
 
 app.listen(PORT, () => {
